+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - MeshElement constructor signature modified to be compatible with code generation.
+- `BBox3.Extend` method is public now
 
 ### Fixed
 - Fixed a bug where `Polyline.Frames` would return inconsistently-oriented transforms.